Do not run `findClosestEnabledDate` for DateRangePicker calendars
