+++ lib/src/views/Calendar/Calendar.tsx
@@ -8,7 +8,6 @@ import { PickerOnChangeFn } from '../../_shared/hooks/useViews';
 import { makeStyles, useTheme } from '@material-ui/core/styles';
 import { DAY_SIZE, DAY_MARGIN } from '../../constants/dimensions';
 import { useGlobalKeyDown, keycode } from '../../_shared/hooks/useKeyDown';
-import { findClosestEnabledDate, DateValidationProps } from '../../_helpers/date-utils';
 import { SlideTransition, SlideDirection, SlideTransitionProps } from './SlideTransition';
 
 export interface ExportedCalendarProps
@@ -30,10 +29,8 @@ export interface ExportedCalendarProps
   loadingIndicator?: JSX.Element;
 }
 
-export interface CalendarProps extends ExportedCalendarProps, DateValidationProps {
+export interface CalendarProps extends ExportedCalendarProps {
   date: MaterialUiPickersDate | MaterialUiPickersDate[];
-  minDate: MaterialUiPickersDate;
-  maxDate: MaterialUiPickersDate;
   isDateDisabled: (day: MaterialUiPickersDate) => boolean;
   slideDirection: SlideDirection;
   currentMonth: MaterialUiPickersDate;
@@ -96,11 +93,7 @@ export const Calendar: React.FC<CalendarProps> = ({
   focusedDay,
   changeFocusedDay,
   onChange,
-  minDate,
-  maxDate,
   slideDirection,
-  disableFuture,
-  disablePast,
   currentMonth,
   renderDay,
   reduceAnimations,
@@ -124,21 +117,6 @@ export const Calendar: React.FC<CalendarProps> = ({
   );
 
   const initialDate = Array.isArray(date) ? date[0] : date;
-  React.useEffect(() => {
-    if (initialDate && isDateDisabled(initialDate)) {
-      const closestEnabledDate = findClosestEnabledDate({
-        utils,
-        date: initialDate,
-        minDate: utils.date(minDate),
-        maxDate: utils.date(maxDate),
-        disablePast: Boolean(disablePast),
-        disableFuture: Boolean(disableFuture),
-        shouldDisableDate: isDateDisabled,
-      });
-
-      handleDaySelect(closestEnabledDate, false);
-    }
-  }, []); // eslint-disable-line
 
   const nowFocusedDay = focusedDay || initialDate;
   useGlobalKeyDown(Boolean(allowKeyboardControl), {
@@ -183,7 +161,7 @@ export const Calendar: React.FC<CalendarProps> = ({
       >
         <div role="grid" style={{ overflow: 'hidden' }}>
           {utils.getWeekArray(currentMonth).map(week => (
-            <div role="row" key={`week-${(week[0] as any)!.toString()}`} className={classes.week}>
+            <div role="row" key={`week-${week[0]}`} className={classes.week}>
               {week.map(day => {
                 const disabled = isDateDisabled(day);
                 const isDayInCurrentMonth = utils.getMonth(day) === currentMonthNumber;
